idがまとまって消えるエラー解決

+++ utils/handleDelete.ts
@@ -1,7 +1,7 @@
 import { db } from "@/lib/firebase";
 import { collection, deleteDoc, getDocs, query, where } from "firebase/firestore";
 
-export const handleDelete = async (id: number, name: string) => {
+export const handleDelete = async (docId: string , name: string , id: number ) => {
   try {
     // 削除前の確認ダイアログ
     const confirmDelete = window.confirm(`本当に製品「${name}」(ID: ${id}) を削除しますか？`);
@@ -11,9 +11,10 @@ export const handleDelete = async (id: number, name: string) => {
     }
 
     const Ref = collection(db, "registr");
-    const req = query(Ref, where("name", "==", name), where("id", "==", id));
+    const req = query(Ref, where("name", "==", name));
     const snapshot = await getDocs(req);
     
+
     if (snapshot.empty) {
       window.alert("その製品は存在しません");
       return;
@@ -21,7 +22,11 @@ export const handleDelete = async (id: number, name: string) => {
 
     // 非同期処理を待ちながら削除
     for (const doc of snapshot.docs) {
+      if( docId === docId ) {
       await deleteDoc(doc.ref);
+      window.alert(`製品「${name}」(ID: ${id})を削除しました`);
+      return;
+      }
     }
 
   } catch (error) {
